+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientUi.kt
@@ -23,4 +23,6 @@ interface ContactPatientUi {
   fun enablePreviousReminderDateStepper()
   fun disableNextReminderDateStepper()
   fun enableNextReminderDateStepper()
+  fun showProgress()
+  fun hideProgress()
 }