+++ app/src/main/java/org/simple/clinic/contactpatient/ContactPatientUpdate.kt
@@ -24,8 +24,8 @@ class ContactPatientUpdate(
       event: ContactPatientEvent
   ): Next<ContactPatientModel, ContactPatientEffect> {
     return when (event) {
-      is PatientProfileLoaded -> next(model.patientProfileLoaded(event.patientProfile))
-      is OverdueAppointmentLoaded -> next(model.overdueAppointmentLoaded(event.overdueAppointment))
+      is PatientProfileLoaded -> patientProfileLoaded(model, event)
+      is OverdueAppointmentLoaded -> overdueAppointmentLoaded(model, event)
       is NormalCallClicked -> directlyCallPatient(model, event)
       is SecureCallClicked -> maskedCallPatient(model, event)
       PatientAgreedToVisitClicked -> dispatch(MarkPatientAsAgreedToVisit(model.appointmentUuid))
@@ -44,6 +44,28 @@ class ContactPatientUpdate(
       PatientMarkedAsAgreedToVisit,
       ReminderSetForAppointment -> dispatch(CloseScreen)
     }
+  }
+
+  private fun overdueAppointmentLoaded(model: ContactPatientModel, event: OverdueAppointmentLoaded): Next<ContactPatientModel, ContactPatientEffect> {
+    val updatedModel = if (!model.hasLoadedPatientProfile) {
+      model.overdueAppointmentLoaded(event.overdueAppointment)
+    } else {
+      model.overdueAppointmentLoaded(event.overdueAppointment)
+          .contactPatientInfoLoaded()
+    }
+
+    return next(updatedModel)
+  }
+
+  private fun patientProfileLoaded(model: ContactPatientModel, event: PatientProfileLoaded): Next<ContactPatientModel, ContactPatientEffect> {
+    val updatedModel = if (!model.hasLoadedAppointment) {
+      model.patientProfileLoaded(event.patientProfile)
+    } else {
+      model.patientProfileLoaded(event.patientProfile)
+          .contactPatientInfoLoaded()
+    }
+
+    return next(updatedModel)
   }
 
   private fun backClicks(model: ContactPatientModel): Next<ContactPatientModel, ContactPatientEffect> {