+++ CHANGELOG.md
@@ -19,7 +19,8 @@
 
 ### Changes
 - Updated translations: `pa-IN`, `hi-IN`, `te-IN`, `kn-IN`, `mr-IN`, `pa-IN`, `te-IN`, `sid-ET`, `kn-IN`, `ta-IN`, `bn-BD`
-- Click on overdue patient to open patient summary  
+- Click on overdue patient to open patient summary
+- Show progress when loading overdue patient contact information  
 - [In Progress: 17 Jun 2021] Overdue list improvements
   - Change overdue list UI
   - Load overdue patients without phone number